requirements.txt
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Flask==2.3.3
google-api-core==2.24.1
google-auth==2.23.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.2
google-cloud-datastore==2.20.2
google-cloud-ndb==2.2.0
google-cloud-secret-manager==2.23.1
googleapis-common-protos==1.68.0
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pymemcache==4.0.0
pytz==2025.1
redis==4.6.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.8.2
six==1.17.0
urllib3==1.26.20
Werkzeug==3.1.3
